pub-catalog/requirements.txt
numpy==1.11.2
boto3==1.4.3
requests==2.12.4
pyyaml
